Propose all conventions, commands and excercises
There are no common git conventions among the colleagues at my company.
The git commit history in most of the repositories is not helpful in
any way.

To change that, this repository shall contain all conventions and
commands you need to maintain a sophisticated commit history. This is
a first overview of all necessary conventions and commands and
exercises I intend to provide.

Consider it as a request for comments