Provide more insights in the practical background of the conventions
The first draft is focused on the pure conventions and hard facts, but
does not provide many insights into the background of those. Practical
examples, real world experiences and the abstract conventions are now
more intertwined with each other.

As the length grow with more examples, the files are moved to dedicated
folders.